Add files via upload
Rules: 
The game of War is a 2 player card game in which each player is given 26 unique cards from a deck of 52. The players each draw 3 cards from the top of their deck at a time and take the average. The player with the higher average takes all cards in play. The first player to 52 cards wins the game. If a player has only 2 remaining cards, they will use the card with the higher value. If both players have the same average, then a "War" will ensue where each player draws another card and the player with the higher card takes all cards in play.

This program uses Object Oriented concepts to simulate the game of war n number of times.